requirements.txt
#  Computer Vision & ML 
opencv-python>=4.8.0
deepface>=0.0.93
tf-keras>=2.16.0
mtcnn>=0.1.1
numpy>=1.24.0

#  Web Framework 
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
websockets>=12.0

#  Database 
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
greenlet>=3.0.0
asyncpg>=0.29.0
alembic>=1.13.0

#  Auth & Security 
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

#  Config & Validation 
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

#  Logging 
python-json-logger>=2.0.7

#  Testing 
pytest>=8.1.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
pytest-cov>=5.0.0

#  Dev Tools 
ruff>=0.4.0
